requirements.txt
scikit-learn
sklearn-deap
numpy
pandas
hdbscan
matplotlib
pyod
PyNomaly
scipy
